requirements.txt
streamlit
pandas
plotly
langchain
langchain-community
google-generativeai